requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
numpy==1.14.2
Pillow==5.1.0
praw==5.4.0
prawcore==0.14.0
pytesseract==0.2.0
requests==2.18.4
update-checker==0.16
urllib3==1.22
